backend/requirements.txt
Flask
Flask-Cors
browser-use
playwright
google-generativeai
